Bulk update up to KanColle-YPS v1.5.8
This version is up to date to v1.5.8 of the master branch of
KanColle-YPS, which (should) fully support the usage of combined fleet.
There were something went wrong during the update process, therefore
this version is completely redone of the update (fixed a typo), plus a
few minor UI improvement.